WASM64 WIP

+++ onnxruntime/wasm/api.h
@@ -145,7 +145,7 @@ ort_tensor_handle_t EMSCRIPTEN_KEEPALIVE OrtCreateTensor(int data_type, void* da
  * @remarks following temporary buffers are allocated during the call. Caller must release the buffers after use by calling OrtFree():
  *           'dims' (for all types of tensor), 'data' (only for string tensor)
  */
-int EMSCRIPTEN_KEEPALIVE OrtGetTensorData(ort_tensor_handle_t tensor, int* data_type, void** data, size_t** dims, size_t* dims_length);
+int EMSCRIPTEN_KEEPALIVE OrtGetTensorData(ort_tensor_handle_t tensor, size_t* data_type, void** data, size_t** dims, size_t* dims_length);
 
 /**
  * release the specified tensor.
